# identy-service/docs/swagger.yaml
basePath: /
definitions:
  http.changePasswordReq:
    properties:
      currentPassword:
        type: string
      newPassword:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
...
  http.lockAccountReq:
    properties:
      locked_until:
        description: LockedUntil is a Unix timestamp (seconds). Null means indefinite
          lock.
        type: integer
    type: object
...
  http.patchAccountReq:
    properties:
      personalInfo:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API для управления идентификацией и доступом SayMo.
  title: SayMo Identity Service API
  version: "1.0"
paths:
...
  /iam/accounts/{accountId}:
    delete:
      description: Мягкое удаление аккаунта
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Soft delete account
      tags:
      - Accounts
    get:
      description: Получение данных аккаунта по ID
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get account by ID
      tags:
      - Accounts
    patch:
      consumes:
      - application/json
      description: Частичное обновление данных аккаунта
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Доля.role и/или personalInfo
        in: body
        name: request
        schema:
          $ref: '#/definitions/http.patchAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Patch account by ID
      tags:
      - Accounts
  /iam/accounts/{accountId}/lock:
    post:
      consumes:
      - application/json
      description: Блокировка аккаунта администратором
      parameters:
      - description: locked_until (Unix timestamp, null for indefinite)
        in: body
        name: request
        schema:
          $ref: '#/definitions/http.lockAccountReq'
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Lock account
      tags:
      - Accounts
  /iam/accounts/{accountId}/password:
    post:
      consumes:
      - application/json
      description: Смена пароля для аккаунта
      parameters:
      - description: Текущий и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.changePasswordReq'
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: password policy violation
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: invalid current password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change password
      tags:
      - Accounts
  /iam/accounts/{accountId}/sessions:
    get:
      description: Получение списка сессий аккаунта
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: List sessions for account
      tags:
      - Accounts
  /iam/accounts/{accountId}/sessions/{sessionId}:
    delete:
      description: Завершение сессии (для админа или владельца)
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Terminate session
      tags:
      - Accounts
  /iam/accounts/{accountId}/unlock:
    post:
      description: Разблокировка аккаунта администратором
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Unlock account
      tags:
      - Accounts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Формат: Bearer {token}.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"